+++ playground/pages/index.vue
@@ -1,9 +1,5 @@
 <script setup lang="ts"></script>
 
 <template>
-  <div class="flex flex-col gap-5">
-    <GButton @click="modal()">drawer</GButton>
-
-    <GModal title="=moz"></GModal>
-  </div>
+  <div class="flex flex-col gap-5"></div>
 </template>