+++ src/runtime/components/G/Modal.vue
@@ -1,5 +1,6 @@
 <script lang="ts" setup>
 import { watch } from "#imports";
+import { GloriousStore } from "../../stores/GloriousStore";
 const props = defineProps({
   modelValue: {
     required: true,
@@ -52,7 +53,7 @@ const emits = defineEmits(["ok", "close", "update:modelValue"]);
 
 const addBlurBackground = (): void => {
   const backgroundBlur = document.createElement("div");
-  backgroundBlur.classList.add("g-scafold-modal-bg-blur");
+  backgroundBlur.classList.add("glorious-scaffold-modal-bg-blur");
   const nuxt: any = document.getElementById("__nuxt");
   nuxt.appendChild(backgroundBlur);
   backgroundBlur.addEventListener("click", () => {
@@ -66,10 +67,21 @@ const addBlurBackground = (): void => {
 watch(
   () => props.modelValue,
   () => {
+    const modals = document.querySelectorAll(".glorious-scaffold-modal");
     if (props.modelValue && props.blur) {
       addBlurBackground();
+      setTimeout(() => {
+        modals.forEach((el: any) => {
+          el.classList.remove("animation");
+        });
+      }, 500);
     } else {
-      const blur: any = document.querySelector(".g-scafold-modal-bg-blur");
+      const blur: any = document.querySelector(
+        ".glorious-scaffold-modal-bg-blur"
+      );
+      modals.forEach((el: any) => {
+        el.classList.add("animation");
+      });
       if (blur !== null) blur.remove();
     }
   }
@@ -78,11 +90,11 @@ watch(
 
 <template>
   <div
-    class="g-scafold-modal"
+    class="glorious-scaffold-modal animation"
     :class="[props.modelValue ? 'open' : 'close', `size-${props.size}`]"
   >
     <!-- header -->
-    <div v-show="props?.title !== ''" class="g-scafold-modal-header">
+    <div v-show="props?.title !== ''" class="glorious-scaffold-modal-header">
       <span class="font-medium">{{ props.title }}</span>
       <GButton
         class="flex justify-center items-center w-6 h-6"
@@ -94,14 +106,14 @@ watch(
       </GButton>
     </div>
     <hr v-show="props.title !== ''" class="my-3" />
-    <!-- end hader -->
+    <!-- end header -->
 
-    <div class="g-scafold-modal-header">
+    <div class="glorious-scaffold-modal-content">
       <slot />
     </div>
 
     <!-- start footer -->
-    <div v-show="props.okBtn !== ''" class="g-scafold-modal-footer">
+    <div v-show="props.okBtn !== ''" class="glorious-scaffold-modal-footer">
       <GButton :loading="props.okBtnLoading" @click="emits('ok')">
         {{ props.okBtn }}
       </GButton>